refactor: create fish job module
This refactor creates a separate module with the completion fish job
abstraction. This makes the code more understandable as we are able to
use separate the nvim-cmp protocol implementation from the details of
controlling a job in Neovim.

I’ve made minor tweaks to when a callback is set in `:send`, but testing
didn’t show any impact of this on results.